+++ src/components/cardItem/article/banner/BannerAvatar.css.ts
