+++ src/components/cardItem/article/banner/BannerText.css.ts
