+++ src/components/cardItem/article/cardBottom/style.css.ts
